requirements.txt
streamlit==1.37.1
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pillow==10.4.0
reportlab==4.2.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
itsdangerous==2.2.0
email-validator==2.1.0.post1
streamlit-drawable-canvas==0.9.3
